Migrate DropRight tests to munit
